requirements.txt
fastapi
numpy
face-recognition
opencv-python
python-jose
passlib
python-dotenv